requirements.txt
fastapi
uvicorn
motor
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
email-validator
aiohttp
slowapi
itsdangerous
certify==2025.7.14